一通り型をつけてコンパイルが通るようにした。動作もちゃんとする。

+++ src/view/TodoItemView.ts
@@ -1,7 +1,23 @@
+import { TodoItemModel } from "../model/TodoItemModel.js";
 import { element } from "./html-util.js";
 
 export class TodoItemView {
-  createElement(todoItem, { onUpdateTodo, onDeleteTodo }) {
+  createElement(
+    todoItem: TodoItemModel,
+    {
+      onUpdateTodo,
+      onDeleteTodo,
+    }: {
+      onUpdateTodo: ({
+        id,
+        completed,
+      }: {
+        id: number;
+        completed: boolean;
+      }) => void;
+      onDeleteTodo: ({ id }: { id: number }) => void;
+    }
+  ) {
     const todoItemElement = todoItem.completed
       ? element`<li><input type="checkbox" class="checkbox" checked>
                                     <s>${todoItem.title}</s>
@@ -11,7 +27,10 @@ export class TodoItemView {
                                     ${todoItem.title}
                                     <button class="delete">x</button>
                                 </li>`;
+    if (!todoItemElement) throw new Error("todoItemElementがnullです");
     const inputCheckboxElement = todoItemElement.querySelector(".checkbox");
+    if (!inputCheckboxElement)
+      throw new Error("inputCheckboxElementがnullです");
     inputCheckboxElement.addEventListener("change", () => {
       onUpdateTodo({
         id: todoItem.id,
@@ -19,6 +38,8 @@ export class TodoItemView {
       });
     });
     const deleteButtonElement = todoItemElement.querySelector(".delete");
+    if (!deleteButtonElement)
+      throw new Error("deleteButtonElementはnullです。");
     deleteButtonElement.addEventListener("click", () => {
       onDeleteTodo({
         id: todoItem.id,
